+++ stories/components/Paragraph.js
@@ -11,7 +11,11 @@ story.add('Default', () => (
     This is an awesome paragraph. It is based on the awesome material-ui Typography component. Its
     default variant is 'body2', and it has a maximal width of 650px. According to awesome studies,
     it is the range where the eye feels the most comfortable reading. In addition to its variant,
-    you can of course use all the material-ui Typography component's props.
+    you can of course use all the{' '}
+    <a href="https://material-ui.com/api/typography/" target="_blank" rel="noreferrer noopener">
+      material-ui Typography component's props
+    </a>
+    .
   </Paragraph>
 ))